// src/enums/MedicationEnum.ts
export enum MedicationScheduleEnum {
  Pending = 'Pending',
  Done = 'Done',
  Not_taken = 'Not taken',
  Cancelled = 'Cancelled',
}

export enum MedicationRequestEnum {
  Pending = 'Pending',
  Scheduled = 'Scheduled',
  In_progress = 'In progress',
  Completed = 'Completed',
  Cancelled = 'Cancelled',
}